chore, refactor: Creado Dao de OperacionPendiente y renombrar su ID

+++ shared/src/commonMain/kotlin/dev/wdona/burnt_out/domain/model/OperacionPendiente.kt
@@ -15,7 +15,7 @@ data class OperacionPendiente(
     val idAccion: Long,
     val tipoAccion: String,
     val tablaAfectada: String,
-    val idAfectado: Int,
+    val idAfectado: Long,
     val datosJson: String,
     val timestampCreacion: Long,
     val sincronizado: Boolean
